+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added get_folder_contents which returns list of the folders/files in the shared folder or one of its sub-folders.
-- Mocked SharedFile object now contains also the type of the file (file or directory).
+- pyndows.get_folder_content listing the content of a folder (non-recursively).
+- File type (isDirectory boolean) within mocked SharedFile object.
 
 ## [3.2.0] - 2020-01-22
 ### Fixed